modified

+++ generator-basic/src/main/java/com/dbr/generator/basic/generator/project/SpringBootProjectGenerator.java
@@ -10,6 +10,10 @@ public class SpringBootProjectGenerator extends ProjectGenerator<SpringBootProje
 
     @Override
     public void execute(SpringBootProjectModel model) throws Exception {
+        super.execute(model);
+
+        if(true)
+            return;
         GeneratorUtil.deleteFile(model.getSpringBootProjectFolder());
         File springBootZipFile = GeneratorUtil.copyUrlToTempFolder(model.getSpringBootTemplateZipUrl(),
                 model.getSpringBootTemplateZipFile());
